+++ example/lib/pages/mini_games/manual_map/game_manual_map.dart
@@ -54,6 +54,7 @@ class GameManualMap extends StatelessWidget {
         cameraConfig: CameraConfig(
           zoom: getZoomFromMaxVisibleTile(context, DungeonMap.tileSize, 20),
         ),
+        showCollisionArea: true,
         interface: KnightInterface(),
         map: DungeonMap.map(),
         backgroundColor: Colors.blueGrey[900]!,