+++ lib/map/spritefusion/world_map_by_spritefusion.dart
@@ -9,19 +9,24 @@ class WorldMapBySpritefusion extends WorldMap {
     WorldMapReader<SpritefusionMap> reader, {
     ValueChanged<Object>? onError,
     double sizeToUpdate = 0,
+    Map<String, SpritefusionObjectBuilder>? objectsBuilder
   }) : super(const []) {
     this.sizeToUpdate = sizeToUpdate;
     _builder = SpritefusionWorldBuilder(
       reader,
       onError: onError,
       sizeToUpdate: sizeToUpdate,
+      objectsBuilder:objectsBuilder,
     );
   }
 
   @override
   Future<void> onLoad() async {
     final map = await _builder.build();
-    layers = map.layers;
+    layers = map.map.layers;
+    gameRef.addAll(map.components ?? []);
     return super.onLoad();
   }
 }
+
+